requirements.txt
# Core dependencies
python-dateutil>=2.8.2
PyJWT>=2.3.0

# Development dependencies
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.12b0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.931

# Optional (for future data export features)
openpyxl>=3.0.9
pandas>=1.3.5
SQLAlchemy>=1.4.32
